+++ lib/views/home/home_screen.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+import 'package:hackathon_supporterz/util/config.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({
+    Key? key,
+    required this.title,
+  }) : super(key: key);
+
+  final String title;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(title),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Text(
+              'Some text in Text widget',
+              style: Config.h1(context),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
